archive/start/requirements.txt
textual==0.9.1
proxmoxer==2.0.1
paramiko==2.12.0
python-dotenv==0.21.0